Proposed change for email sender info in com_contact to better combat spam
The default behaviour of the contacts component is to use the site's
name and email address as the sender of all contact email that gets
sent via com_contact forms with a reply-to set to the actual sender. If
we leave it like this, then antispam filters cannot be trained to
distinguish spam from regular email. The proposed change is to use the
actual sender's name and email address in all the email's headers (from
& reply-to).