Configuration update.
A configuration implementation with multiple resolvers. With there being
a JConfigResolverInterface it will allow people to create their own
where the may want to load from a internal url etc.

Allow for .env.php in configuration directory to load environmental
variables such as J_DEBUG and J_HOST. This means you can create a git
repository and not have to commit the configuration.

This is untested and needs tweaking but I'm just trying to get people
started on improving the configuration system. The current one makes it
really hard for developers to use git tracked projects using Joomla to
have separate configuration. I have tried to keep it as much in tact as
possible with the JFactory::getConfig method.