Add ability to override the renderer
In Joomla! if you want to override something you have to put it in the active template directory (e.g.: the layouts goes in (active template)/html/layouts).
This PR introduces the ability to override the whole rendering process in the template by copying (and modifying per your preferences) the /libraries/joomla/document directory.

It is shouldn't rise any backward compatibility issues (except that the directory [active template]/document might be in use?).

What this might provide?
Flexibility on the way that the template renders (the inner logic) the data. e.g. someone may use echo $this->getBuffer('modules', 'menu', array('menu', 'html5')); and remove the <jdoc..> parser.

Why?
Because this is one area that if you want to override and use your own logic you have to do it through a plugin (this is way easier and cleaner)