Added missing capability to remove attachment(s)
It might be wished (at least by me) to remove email a specific email attachment or all attachments at once.

I my case i send an email to one user and a confirmation email to another. While one user requires the attachments the other does not. The processing flow requires the attachments to first be added and then to be removed before sending the second mail.
While PHPmailer itself provides the capbility to clear attachments the wrapper class `JMail` does not (yet).