Make blog item image linked to article
I believe that it should either be an option or always setup that the image also links to the article just like the title and readmore.

(This is my first try at contributing to Joomla. Please tell help me if I am doing this wrong. I am not a coder, but was able to get this to work for my client's site by copy/pasting similar code.)