make intro image link to article
I believe that it should either be an option or always setup that the image also links to the article just like the title and readmore.  In the blog layout I watch my clients always rolling over the intro image first to click on it to see the full article.  I do this to all my sites as it seems to be a better UX.

(This is my 2nd :-) try at contributing to Joomla. Please tell help me if I am doing this wrong. I am not a coder, but was able to get this to work for my client's site by copy/pasting similar code.)