Extended JImageHelper
I found that with the use of base64-encoded data streams of image files it should be possible to integrate the creation of such streams and that JImageHelper is the best place to request it from. So i added two new methods - toBase64 and fromBase64. While the former method generates a ready to use base64-encoded data stream that can be used with an <img/> element's src-attribute the latter takes such a stream and write it into an image file.