Status of modules : missing pending and expired
Fix to Brian request : https://github.com/joomla/joomla-cms/issues/4414
<pre>
Create a module with an end publishing date in the PAST. In the module list there is no visible change to its publishing status. It is still green. If you check an expired article you will see that the icon in the article list is orange not green to indicate it has expired.
</pre>