Remove close tag for an unopened span element.
This is probably a close tag from the default.php list view where the id is enclosed in a span element with a title set to the lft and rgt of the item. The starting tag was removed but the closing tag was still there. Removing the close tag results in a cleaner and valid html code.