This PR fixes a regression introduced by #4221
#4221 was incorrectly using str_replace() to remove the default language
code when needed. As str_replace acts globally on the subject, the
language code was removed also out of context (e.g. when part of an
alias).

With this PR I use instead preg_replace with the limit of one
substitution. due to the fact that the /language-code string to be
replaced appears just after the host name and cannot be part of it (due
to the/), this should guarantee the correct replacements of the language
code only.

I took the occasion to make a bit of clean-up by aggregating some
redundantly duplicated code