review the installation/configuration.php-dist
See: https://github.com/joomla/joomla-cms/issues/4498

This is a review of the installation/configuration.php-dist file. I have add some comments move the default dbtype to mysqli and outcommend the root_user (<-- or is this realy needed?)

I don't know if this need any testing but it can be tested is the following:

1. download a clean staging
2. extract
3. move this file to the joomla root
4. change the needed settings
5. install the database manualy
6. make sure all works as if you use the installer.

Thanks @robations for pointing it out.