JHtmlBootstrap::tooltip() hide vs. hideme
It appears the `hide` event is not fired when triggering an element to show the tooltip. However, the `hideme` event is fired. This can easily be tested with the following code:
```
JHtml::_('bootstrap.tooltip', '.hasStickyTooltip', array(
    'onShow'   => 'function ( e ) {'
        .'alert("onShow executed")'.
    '}',
    'onShown'  => 'function ( e ) {'
        .'alert("onShown executed")'.
    '}',
    'onHide'   => 'function ( e ) {'
        .'alert("onHide executed")'.
    '}',
    'onHideMe' => 'function ( e ) {'
        .'alert("onHideMe executed")'.
    '}',
    'onHidden' => 'function ( e ) {'
        .'alert("onHidden executed")'.
    '}'
));
```
This proves the message `onHide executed` never appears. To get hands on the event fired before the widget is getting hidden i added the corresponding listener `onHideMe` to the configuration array. With this listener in place the message `onHideMe executed` is shown and the corresponding event is handled by the plugin.