Move user field to bs modal
Same as #4513 but for User

User select field (field type user) is using mootools modal.
Without affecting in anyway the logic of the field we can use jQuery and bootstrap.

B/C
Should be 100% compatible

Test:
Apply the patch
Go to administrator -> Banners -> create a new banner
and check the functionality of created by in publishing tab
Everything should be ok