Move content history to bs modal
Same as #4513 and 4514 but for content history

Content history is currently using mootools modal.
Without affecting in anyway the logic of the field we can use jQuery and bootstrap.

B/C
Should be 100% compatible

Test:
Apply the patch
Go to administrator -> Banners -> create a new banner
change the alternative text and save it again
repeat the last part for few more times

press the versions button
Everything should be ok