changes reverted
Travis reported many errors. But it appeard not of them relate to this file. However, i think the compression algorithm used for the original version of this file comes from a different compressor. So mabe the compressed versions are created when the release is prepare for deployment? Anyway, i leave this file like it was before i applied my change.