[imp] Making sure Components menu is not case sensitive. Fixes #4889
In one of the previous Joomla! 3 major versions, 3.0, 3.1 or 3.2, I can't remember which, the default ordering for the Components menu was changed to be case sensitive. From that moment, extensions that started with a lowercase letter in their name would be listed at the bottom of the Components menu. This pull request reverts that change back, so the Components menu is truly in alphabetical order again.