batch exec queries instead as each as single query
The reodering in the JTable executes any reordering statement as a single query. If you have a lot of articles in one category it can easily add hunderts of query this results in a large response time and can be result in timeouts.

This is hard to find out, because of the redirect after a save the debug shows only the executed queries from the last request and not from the save request.

This patch implements for the mysqli DB driver a executeBatch command that can handle an array of queries, concats the queries and execute it with the  mysqli_multi_query function. This is significant faster, I had on my local env. 1:15 before and 0:02 after the patch. (just watch the videos)