Fix warning in installer.php when tmp dir deleted
If the expected manifest XML files/dirs are missing, the findManifest() method fails with:

    [ErrorException] array_merge(): Argument #1 is not an array

because `$parentXmlfiles` and/or `$allXmlFiles` are false, rather than an array.

Caveat - although this fixed the problem for me, I don't fully understand the context of this code so perhaps there is a deeper problem? I was using the [Joomlatools Composer Installer](https://github.com/joomlatools/joomla-composer) to install a module and had deleted the contents of the tmp directory before this happened. Not sure why the tmp dir would be referenced to see if a module/component/plugin is installed, obviously the tmp dir will get cleared out periodically.