add a line between two options for better UI
there is no line between the option to show/hide featured articles and the amount of articles to show in this module.
Without the line people (read: me) can read it as: How many featured articles can be shown. Therefor I was looking for the field to set the amount of other articles I can show. 

With the line between the option show/hide featured articles and the field amount people (read: me) have better understanding of the meaning of amount.