Update tags.php
When using getTagItemsQuery() in JHelperTags you can specify whether or not to include_children. Setting this to 'true' ends up in a call to getTagTreeArray() (line 511), whereby an attempt is made to instantiate an JTable object (line 642). However, as the include path isn't updated to include the necessary tables folder this fails, and calling the isLeaf() method (line 644) on the supposed JTable results in:

Fatal error: Call to a member function isLeaf() on a non-object in /home/account/public_html/libraries/cms/helper/tags.php on line 647

I think the fix is just to update the include path before the instantiation - tried it on my J! 3.3.6 install and the error disappears.