Removed index.html, prepared for installation pack
As the index.html files have bin removed should those files be removed from the manifest files.
Not urgent since it only gives an error when trying to (re-)install with an installation package.
I tested uninstall and discover and got no messages. Uninstall removed an existing index.html 
without any message even after I removed it from the manifest file.

In addition I prepared for an installation package.
- added method="upgrade"
- changed the language to be loaded from a language folder language/en-GB/ to get a translated description after installation.

If this is accepted all manifests should be edited at least when they get edited anyway. 
I choose this as an example since it is a new plugin, I think disabled by default and could be uninstalled in a lot of Joomla use cases.