Fix `$hits_class` generation
This PR fixes the `$hits_class` generation.

### current behavior

every time the class is `info`

### expected behavior

class should change for the hits. (red/yellow/blue)

### Open questions

Why we move it to red if the hits is more than 100? All my artikles have more that 100 hits (thats in that list) Maybe we need to increase the vaules?

Example
10.000 -> `important`
1.000 -> `warning`
all others -> `info`

code:
```php
$hits_class = ($hits >= 10000 ? 'important' : ($hits >= 1000 ? 'warning' : 'info'));
```

### Orginaly reported in german

http://www.joomlaportal.de/joomla-3-x-erweiterungen-module/316386-kleiner-bug-mod_popular.html