Update form.php
In case of a validation failure, the (error) message is not displayed. This is due to the fact that an Exception object is pushed onto the JForm::errors array, while this must be a text string apparently.

Line 1212
array_push($this->errors, $valid);

has now been changed into
array_push($this->errors, $valid->getMessage());

Which seems to solve the problem.