Missing string in front end. Fixes #6206. Fixes #6210
As of Joomla 3.4 alias incrementing was implemented making it possible for users to more easily create articles with the same name. If an article named "joomla-tutorial" already existed and someone creates another one, the alias will automatically become "joomla-tutorial-1" and so on. In the backend when submitting an already existing alias it says "Warning Alias already existed so a number was added at the end. You can re-edit the article to customise the alias." which is an informative warning. However when submitting an article on the frontend it says "Warning COM_CONTENT_SAVE_WARNING". The language constant does not display the intended message for frontend users, but does for backend users.