Fix option name must not be case sensitive
When trying to get a component's parameters passing in the component name in uppercase format that component won't be found and cause the error `JLIB_APPLICATION_ERROR_COMPONENT_NOT_FOUND` be thrown, where as when passing the same name in lower case format everything runs well.
I think case sensitivity should play absolutely not a role at this important place to ensure the component (parameters) will be found no matter what case the option string is formatted like. Thus i added every method in this class which get the component name passed in a case fix.