Fix for languagefilter problem
# Executive summary
This is the fix for a with 3.4 invented bug. For better understanding it comes with a some more refactoring and removing duplicated code.

# Backwards compatibility
Full b/c, no problems are expected

# Translation impact
none

# Testing instructions
You need a multi language site to test, with SEF enabled. I have

de == default language
en == other language

## Wrong behaviour

1. access the site without lang code e.g. http://mysite.de

-> you should see the site in default language

2. change the lang code to en

-> you should see the site in english language

3. access the site without lang code e.g. http://mysite.de again

-> you’ll see the site in default language but it should be in english

## Right behaviour

1. access the site without lang code e.g. http://mysite.de

-> you should see the site in default language

2. change the lang code to en

-> you should see the site in english language

3. access the site without lang code e.g. http://mysite.de again

-> you should see the site in english language

Test if switching language works as expected, also test with the setting Remove URL Language Code to yes. It makes also a difference if the site is installed in a subdir or not.