Dont show database prefix when throwing Exception
If an SQL error occur, database driver throw a runtimeException containg
the SQL with prefix replaced by table prefix.

This fix show in exception message the sql before the string identifier
was replaced.

I think this is a tiny security improvment, by that we can't obtain the
secret prefix.

Unfortunately, im unable to test on many drivers, moreover, PDO handle
the error differently so i actually not modified this driver.