CLEANING-JDatabaseQuery
Moved the JDatabaseQueryElement to is own file, because there is nothing gained by having multiple classes in a single file and it violates the naming conventions.

I also make the constructor element an optional null value, this isn't a BC break, because it was required before, so anyone using it will be sending a value.

Making it optional in the constructor allow us to remove all the if/else clauses in the JDatabaseQuery class.